Added Exception handling

+++ src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -65,6 +65,10 @@ public class TodoController {
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+		
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
